Move arch-specific type definitions into arch-specific modules
